refactor(routes): extract jellyfin lifecycle routes
